fyp/requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
sentence-transformers==2.2.2
torch>=1.9.0
numpy>=1.21.0